backend/requirements.txt
# FastAPI and Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# Payment Processors
stripe==7.8.0
paypalrestsdk==1.13.1
braintree==4.23.0

# Facebook Integration
requests==2.31.0

# Email & SMS
sendgrid==6.11.0
twilio==8.11.1

# AI & ML (Optional)
openai==1.3.7
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
python-dateutil==2.8.2

# Logging & Monitoring
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
